+++ extensions/databinding/src/test/java/me/drakeet/multitype/extensions/databinding/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.camnter.multitype.databinding;
+package me.drakeet.multitype.extensions.databinding;
 
 import org.junit.Test;
