+++ sample/src/main/res/layout/activity_binding_normal_list.xml
@@ -28,7 +28,7 @@
 
         <variable
             name="collaborator"
-            type="com.camnter.multitype.databinding.sample.normal.vm.NormalCollaborator"/>
+            type="com.camnter.multitype.databinding.sample.normal.collaborator.NormalCollaborator"/>
 
     </data>
